Don't error out if default-directory doesn't exist.
Changed editorconfig-get-properties to set default-directory to "/"
before calling create-process.  Create-process tries to set the current
directory from default-directory and will error out if this directory
doesn't exist.  This was causing editorconfig-get-properties to fail
when visiting a file in a directory that doesn't exist (yet).

Also use with-temp-buffer to create and manage the temporary buffer used
by editorconfig-get-properties instead of duplicating that functionaily
by hand.