Avoid asking coding system when not specified
Why ?

In short: Emacs always asks which charset to use when
editorconfig-mode is enabled, .editorconfig does not exist and trying
to save files with non-ascii characters.  This is an annoying behavior
so stop it.

Calling `set-buffer-file-coding-system` sets the "explicit" flag, and
it is used just before saving files.
When this flag is set but the charset is `undecided`, Emacs asks
users what charset to use for saving files with non-ascii
characters (when this flag is not set Emacs do not ask and decide
to use something automatically, `utf-8` for my environment).

This typically happens when editing files in repositories without
.editorconfig, and it is a bit annoying behavior.
This change makes this plugin not call `set-buffer-file-coding-system`
at all when neither `charset` nor `end_of_line` is provided, thus
Emacs can decide charset automatically when saving files in
repositories without .editorconfig files.