Add php-mode to indentation-alist and add a short note
About `php-mode-coding-style'.

They basically do this at mode startup: "if php-mode-coding-style is
'pear' (which it is by default), add a c-style called 'pear' and
switch to it", and that style contains a c-basic-offset value. But
since we're hooking later than that, we're welcome to override
c-basic-offset without any risk.

Also, php-mode suggests manually typing "nxml-mode" and "js2-mode" and
"web-mode" and then "php-mode" etc, to switch editing styles if you
have a document with multiple types of content (which is common in
.php files), and every time you type php-mode, it runs that style
injection again and resets c-basic-offset.
However, editorconfig runs as a hook when the major mode changes,
so it still re-applies the correct offset every time that happens.

See #116 for more details.