Always handle "unset" values as non-present
According to [the editorconfig specification][1],

> For any pair, a value of `unset` removes the effect of that pair, even if it
> has been set before. For example, add `indent_size = unset` to undefine the
> `indent_size` pair (and use editor defaults).

Yet, the test suite for editorconfig *cores* [indicates][2] that an editorconfig
shouldn't ignore values of "unset", and should return them as-is.

Therefore it's up to editorconfig *plugins* to discard values of "unset" at the
point of use.

Before this commit, the only place in the code where the value "unset" was
checked for was in `editorconfig-set-major-mode-from-ext`, whereas in other
places, such as `editorconfig-set-indentation`, it was not checked.

After this commit, values of "unset" are immediately removed from the properties
hashtable before it is used to do anything in `editorconfig-apply`.  This
ensures that properties with value "unset" will be treated indistinguishably
from unspecified properties.

This fixes the behavior shown in this gist:

  https://gist.github.com/kini/560673dbff4b8e42bb40a82a8ca6e5b1

[1]: https://editorconfig-specification.readthedocs.io/en/latest/#supported-pairs
[2]: https://github.com/editorconfig/editorconfig-core-test/issues/12