editorconfig.el: skip special-mode buffers when applying (#246)
Some buffers get confused if you keep setting their mode, one example
is git-rebase-mode. This is an example buffer of a class of
special-mode buffers which are usually (but not always) non-file
buffers. Either way these buffers are usually special purpose
interactive buffers that will only get confused if editorconfig tries
to mess around with them.