bin/editorconfig-el: Fix test failure with snapshot (#384)
The tests don't like the "missing lexical-binding" warnings we get
from Emacs-31.  It's actually difficult to fix those warnings
because the `lexical-binding` cookie has to be on the first line
or on the second line but only if that second line starts with ";",
and I don't know of a way to satisfy those constraints are still have
a runnable shell script.

There are various possible workarounds, such as cutting the first few
lines before we pass the content of `$0` to Emacs, but it seemed
simpler to just stamp out the function that emits the warning.