New variable `editorconfig-indent-size-vars` (#383)
This commit aligns indentation var rules with those of Emacs-30

This introduces a new var `editorconfig-indent-size-vars` to determine
how to set which variable for the current major mode.
For modes which don't set this var, the default value
still looks up `editorconfig-indentation-alist` but the
documented format of that var is changed in two ways:

- For `(MODE . FUNCTION)`, FUNCTION should now return the settings
  as a list rather than apply the settings directly.
- For `(MODE . INDENT-SPEC-LIST)`, the elements in the list cannot
  be cons cells any more, only variable names.

The old format elements are still supported but a warning
message is emitted when we encounter them.

In addition, if no entry is found in `editorconfig-indentation-alist`
the default now tries to guess an appropriate variable name, by
trying out a few common naming conventions.

* editorconfig.el (editorconfig-indent-size-vars): New var.
(editorconfig--default-indent-size-function): New function.
(editorconfig--get-indentation): New function.
(editorconfig-set-indentation): Rewrite using it.
(editorconfig--get-indentation-web-mode)
(editorconfig--get-indentation-nxml-mode): New functions.
(editorconfig--get-indentation-(python|latex|lisp)-mode):
Rename from `editorconfig-set-indentation-*-mode` and make them return
a list of settings instead of setting the vars directly.
(editorconfig-indentation-alist): Remove entries for modes that follow
a common enough naming convention that
`editorconfig--default-indent-size-function` is able to guess the
relevant var's name.
Use the new `editorconfig--get-*` functions.
Adjust docstring to the new format.

Co-authored-by: Hong Xu <hong@topbug.net>